OCR/Tesseract/requirements.txt
opencv-python==4.5.5.64
pytesseract~=0.3.10
Pillow~=9.5.0
easyocr~=1.6.2
boto3==1.26.142
botocore==1.29.142
awslambdaric==2.0.4
av==10.0.0
imageio==2.29.0